fix ratio decimals
